moved read rows tests in test_client
